x/shitchain/types/message_create_thread_test.go
package types

import (
	"testing"

	"github.com/bonedaddy/shitchain/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgCreateThread_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateThread
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateThread{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateThread{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
